[TIMOB-17876] 64 bit safe code migration
Turn on GCC_WARN_64_TO_32_BIT_CONVERSION flag and fix code for the same

+++ iphone/Classes/AFOpenFlow/AFItemView.m
@@ -55,11 +55,11 @@
 	self.frame = CGRectMake(0, 0, newImage.size.width, newImage.size.height);
 }
 
-- (int)number {
+- (NSUInteger)number {
 	return number;
 }
 
-- (void)setNumber:(int)newNumber {
+- (void)setNumber:(NSUInteger)newNumber {
 	horizontalPosition = COVER_SPACING * newNumber;
 	number = newNumber;
 }
