+++ iphone/Classes/TiMediaVideoPlayerProxy.h
@@ -24,7 +24,7 @@
 	NSMutableArray *views;
 	TiFile *tempFile;
 	KrollCallback *thumbnailCallback;
-	int callbackRequestCount;
+	NSUInteger callbackRequestCount;
 	
 	NSMutableDictionary* loadProperties; // Used to set properties when the player is created
 	BOOL sizeSet;
